requirements/test.txt
pytest-django==3.9.0
factory-boy==2.12.0
mock==4.0.2
pytest==5.4.2
httpretty==1.0.2